Tidy up (#232)
