+++ src/Factory/AbstractMethodInvocationFactory.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace webignition\BasilCompilableSource\Factory;
+
+abstract class AbstractMethodInvocationFactory
+{
+    protected ArgumentFactory $argumentFactory;
+
+    public function __construct(ArgumentFactory $argumentFactory)
+    {
+        $this->argumentFactory = $argumentFactory;
+    }
+}
