+++ src/Factory/MethodInvocationFactory.php
@@ -7,15 +7,8 @@ namespace webignition\BasilCompilableSource\Factory;
 use webignition\BasilCompilableSource\MethodInvocation\MethodInvocation;
 use webignition\BasilCompilableSource\MethodInvocation\MethodInvocationInterface;
 
-class MethodInvocationFactory
+class MethodInvocationFactory extends AbstractMethodInvocationFactory
 {
-    private ArgumentFactory $argumentFactory;
-
-    public function __construct(ArgumentFactory $argumentFactory)
-    {
-        $this->argumentFactory = $argumentFactory;
-    }
-
     public static function createFactory(): self
     {
         return new MethodInvocationFactory(
