+++ src/Factory/ObjectConstructorFactory.php
@@ -7,15 +7,8 @@ namespace webignition\BasilCompilableSource\Factory;
 use webignition\BasilCompilableSource\ClassName;
 use webignition\BasilCompilableSource\MethodInvocation\ObjectConstructor;
 
-class ObjectConstructorFactory
+class ObjectConstructorFactory extends AbstractMethodInvocationFactory
 {
-    private ArgumentFactory $argumentFactory;
-
-    public function __construct(ArgumentFactory $argumentFactory)
-    {
-        $this->argumentFactory = $argumentFactory;
-    }
-
     public static function createFactory(): self
     {
         return new ObjectConstructorFactory(
