+++ tests/Unit/Factory/StaticObjectMethodInvocationFactoryTest.php
@@ -4,13 +4,12 @@ declare(strict_types=1);
 
 namespace webignition\BasilCompilableSource\Tests\Unit\Factory;
 
-use webignition\BasilCompilableSource\Expression\LiteralExpression;
 use webignition\BasilCompilableSource\Factory\StaticObjectMethodInvocationFactory;
 use webignition\BasilCompilableSource\MethodInvocation\MethodInvocation;
 use webignition\BasilCompilableSource\MethodInvocation\StaticObjectMethodInvocation;
 use webignition\BasilCompilableSource\StaticObject;
 
-class StaticObjectMethodInvocationFactoryTest extends \PHPUnit\Framework\TestCase
+class StaticObjectMethodInvocationFactoryTest extends AbstractMethodInvocationFactoryTest
 {
     private StaticObjectMethodInvocationFactory $factory;
 
@@ -51,28 +50,11 @@ class StaticObjectMethodInvocationFactoryTest extends \PHPUnit\Framework\TestCas
             'non-empty' => [
                 'staticObject' => $staticObject,
                 'methodName' => 'methodName',
-                'arguments' => [
-                    100,
-                    M_PI,
-                    'string without single quotes',
-                    'string with \'single\' quotes',
-                    true,
-                    false,
-                    new \stdClass(),
-                    new StaticObject('self'),
-                ],
+                'arguments' => $this->getArguments(),
                 'expectedInvocation' => new StaticObjectMethodInvocation(
                     $staticObject,
                     'methodName',
-                    [
-                        new LiteralExpression('100'),
-                        new LiteralExpression((string) M_PI),
-                        new LiteralExpression('\'string without single quotes\''),
-                        new LiteralExpression('\'string with \\\'single\\\' quotes\''),
-                        new LiteralExpression('true'),
-                        new LiteralExpression('false'),
-                        new StaticObject('self'),
-                    ]
+                    $this->getExpectedArguments()
                 ),
             ],
         ];
